Move some windows into screens folder
