+++ packages/editor/src/routes.tsx
@@ -10,11 +10,11 @@ import MagickPageLayout from './layouts/MagickPageLayout/MagickPageLayout'
 import MainLayout from './layouts/MainLayout/MainLayout'
 import HomeScreen from './screens/HomeScreen/HomeScreen'
 import Magick from './screens/Magick/Magick'
-import AgentManagerWindow from './windows/agents/AgentManagerWindow'
-import DocumentWindow from './windows/DocumentWindow'
-import EventWindow from './windows/EventWindow'
-import RequestWindow from './windows/RequestWindow'
-import SettingsWindow from './windows/settings/SettingsWindow'
+import AgentManagerWindow from './screens/agents/AgentManagerWindow'
+import DocumentWindow from './screens/DocumentWindow'
+import EventWindow from './screens/EventWindow'
+import RequestWindow from './screens/RequestWindow'
+import SettingsWindow from './screens/settings/SettingsWindow'
 // import ProjectWindow from './windows/ProjectWindow';
 import './App.css'
 import './design-globals/design-globals.css'
