+++ packages/editor/src/screens/DocumentWindow/index.module.scss
