+++ packages/editor/src/screens/EventWindow/index.module.scss
