+++ packages/editor/src/screens/agents/AgentWindow/Accordion.tsx
