+++ packages/editor/src/screens/agents/AgentWindow/AgentDetails.tsx
