+++ packages/editor/src/screens/agents/AgentWindow/index.module.scss
