+++ packages/editor/src/screens/agents/AgentWindow/spellValidator.ts
